+++ R/PipeOpFPCA.R
@@ -24,8 +24,6 @@
 #'
 #' @export
 #' @examples
-#' library(mlr3pipelines)
-#'
 #' task = tsk("fuel")
 #' po_fpca = po("fda.fpca", n_components = 3L)
 #' task_fpca = po_fpca$train(list(task))[[1L]]