+++ tests/testthat/test_tf.R
@@ -4,9 +4,9 @@ test_that("tf does not support NAs", {
   # in which case we then have to adress this case in the PipeOps
   # https://github.com/tidyfun/tf/issues/33
   # Currently, NA functions are dropped
-  d = data.frame(time = 1, value = NA_real_, id = "1")
+  d = data.frame(time = 1, value = NA_real_, id = "1", stringsAsFactors = FALSE)
   x = invisible(tf::tfd(d, arg = "time", value = "value", id = "id"))
-  expect_true(length(x) == 0)
+  expect_length(x, 0L)
 })
 
 test_that("pofu has no surprises: irreg", {