+++ NEWS.md
@@ -1,5 +1,8 @@
 # dev
 
+* New PipeOp `PipeOpFDACor`
+* mlr3fda now depends on mlr3pipelines instead of importing it
+
 # mlr3fda 0.1.1
 
 * Fix CRAN issues